ActualizacionPOS

+++ backend/svir-api/src/main/java/com/svir/api/dto/ingrediente/IngredienteResponse.java
@@ -3,6 +3,8 @@ package com.svir.api.dto.ingrediente;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class IngredienteResponse {
@@ -10,7 +12,7 @@ public class IngredienteResponse {
     private Long id;
     private String nombre;
     private String unidadMedida;
-    private Integer stock;
-    private Integer stockMinimo;
+    private BigDecimal stock;
+    private BigDecimal stockMinimo;
     private Boolean activo;
 }
